move model functions into model API
