+++ tests/api/test_model.py
@@ -5,6 +5,7 @@ import numpy as np
 import pytest
 
 from lightning_pose.api import Model
+from lightning_pose.api.model import get_model_class
 from tests.fetch_test_data import fetch_test_data_if_needed
 
 
@@ -262,3 +263,70 @@ def test_predict_frame_bbox_clipping(tmp_path, request):
     assert np.all(kp[:, 1] >= 0)
     assert np.all(kp[:, 0] <= 640 + 1)
     assert np.all(kp[:, 1] <= 480 + 1)
+
+
+class TestGetModelClass:
+    """Test the get_model_class function."""
+
+    def test_get_model_class_supervised_regression(self):
+        """Returns RegressionTracker for supervised regression."""
+        from lightning_pose.models import RegressionTracker
+        assert get_model_class('regression', semi_supervised=False) is RegressionTracker
+
+    def test_get_model_class_supervised_heatmap(self):
+        """Returns HeatmapTracker for supervised heatmap."""
+        from lightning_pose.models import HeatmapTracker
+        assert get_model_class('heatmap', semi_supervised=False) is HeatmapTracker
+
+    def test_get_model_class_supervised_heatmap_mhcrnn(self):
+        """Returns HeatmapTrackerMHCRNN for supervised heatmap_mhcrnn."""
+        from lightning_pose.models import HeatmapTrackerMHCRNN
+        assert get_model_class('heatmap_mhcrnn', semi_supervised=False) is HeatmapTrackerMHCRNN
+
+    def test_get_model_class_supervised_heatmap_multiview_transformer(self):
+        """Returns HeatmapTrackerMultiviewTransformer for supervised multiview transformer."""
+        from lightning_pose.models import HeatmapTrackerMultiviewTransformer
+        assert (
+            get_model_class('heatmap_multiview_transformer', semi_supervised=False)
+            is HeatmapTrackerMultiviewTransformer
+        )
+
+    def test_get_model_class_supervised_raises_for_unknown(self):
+        """Raises NotImplementedError for an unrecognised supervised model_type."""
+        with pytest.raises(NotImplementedError, match='invalid model_type for a fully supervised'):
+            get_model_class('unknown_type', semi_supervised=False)
+
+    def test_get_model_class_semi_supervised_regression(self):
+        """Returns SemiSupervisedRegressionTracker for semi-supervised regression."""
+        from lightning_pose.models import SemiSupervisedRegressionTracker
+        assert (
+            get_model_class('regression', semi_supervised=True) is SemiSupervisedRegressionTracker
+        )
+
+    def test_get_model_class_semi_supervised_heatmap(self):
+        """Returns SemiSupervisedHeatmapTracker for semi-supervised heatmap."""
+        from lightning_pose.models import SemiSupervisedHeatmapTracker
+        assert get_model_class('heatmap', semi_supervised=True) is SemiSupervisedHeatmapTracker
+
+    def test_get_model_class_semi_supervised_heatmap_mhcrnn(self):
+        """Returns SemiSupervisedHeatmapTrackerMHCRNN for semi-supervised heatmap_mhcrnn."""
+        from lightning_pose.models import SemiSupervisedHeatmapTrackerMHCRNN
+        assert (
+            get_model_class('heatmap_mhcrnn', semi_supervised=True)
+            is SemiSupervisedHeatmapTrackerMHCRNN
+        )
+
+    def test_get_model_class_semi_supervised_heatmap_multiview_transformer(self):
+        """Returns SemiSupervisedHeatmapTrackerMultiviewTransformer for semi-supervised variant."""
+        from lightning_pose.models import SemiSupervisedHeatmapTrackerMultiviewTransformer
+        assert (
+            get_model_class('heatmap_multiview_transformer', semi_supervised=True)
+            is SemiSupervisedHeatmapTrackerMultiviewTransformer
+        )
+
+    def test_get_model_class_semi_supervised_raises_for_unknown(self):
+        """Raises NotImplementedError for an unrecognised semi-supervised model_type."""
+        with pytest.raises(
+            NotImplementedError, match='invalid model_type for a semi-supervised',
+        ):
+            get_model_class('unknown_type', semi_supervised=True)
