+++ tests/utils/test_predictions.py
@@ -4,82 +4,11 @@ import copy
 import gc
 
 import lightning.pytorch as pl
-import pytest
 import torch
 from lightning.pytorch.callbacks import ModelCheckpoint
 
-from lightning_pose.utils.predictions import (
-    get_model_class,
-    predict_dataset,
-)
+from lightning_pose.utils.predictions import predict_dataset
 from lightning_pose.utils.scripts import get_loss_factories, get_model
-
-
-class TestGetModelClass:
-    """Test the get_model_class function."""
-
-    def test_get_model_class_supervised_regression(self):
-        """Returns RegressionTracker for supervised regression."""
-        from lightning_pose.models import RegressionTracker
-        assert get_model_class('regression', semi_supervised=False) is RegressionTracker
-
-    def test_get_model_class_supervised_heatmap(self):
-        """Returns HeatmapTracker for supervised heatmap."""
-        from lightning_pose.models import HeatmapTracker
-        assert get_model_class('heatmap', semi_supervised=False) is HeatmapTracker
-
-    def test_get_model_class_supervised_heatmap_mhcrnn(self):
-        """Returns HeatmapTrackerMHCRNN for supervised heatmap_mhcrnn."""
-        from lightning_pose.models import HeatmapTrackerMHCRNN
-        assert get_model_class('heatmap_mhcrnn', semi_supervised=False) is HeatmapTrackerMHCRNN
-
-    def test_get_model_class_supervised_heatmap_multiview_transformer(self):
-        """Returns HeatmapTrackerMultiviewTransformer for supervised multiview transformer."""
-        from lightning_pose.models import HeatmapTrackerMultiviewTransformer
-        assert (
-            get_model_class('heatmap_multiview_transformer', semi_supervised=False)
-            is HeatmapTrackerMultiviewTransformer
-        )
-
-    def test_get_model_class_supervised_raises_for_unknown(self):
-        """Raises NotImplementedError for an unrecognised supervised model_type."""
-        with pytest.raises(NotImplementedError, match='invalid model_type for a fully supervised'):
-            get_model_class('unknown_type', semi_supervised=False)
-
-    def test_get_model_class_semi_supervised_regression(self):
-        """Returns SemiSupervisedRegressionTracker for semi-supervised regression."""
-        from lightning_pose.models import SemiSupervisedRegressionTracker
-        assert (
-            get_model_class('regression', semi_supervised=True) is SemiSupervisedRegressionTracker
-        )
-
-    def test_get_model_class_semi_supervised_heatmap(self):
-        """Returns SemiSupervisedHeatmapTracker for semi-supervised heatmap."""
-        from lightning_pose.models import SemiSupervisedHeatmapTracker
-        assert get_model_class('heatmap', semi_supervised=True) is SemiSupervisedHeatmapTracker
-
-    def test_get_model_class_semi_supervised_heatmap_mhcrnn(self):
-        """Returns SemiSupervisedHeatmapTrackerMHCRNN for semi-supervised heatmap_mhcrnn."""
-        from lightning_pose.models import SemiSupervisedHeatmapTrackerMHCRNN
-        assert (
-            get_model_class('heatmap_mhcrnn', semi_supervised=True)
-            is SemiSupervisedHeatmapTrackerMHCRNN
-        )
-
-    def test_get_model_class_semi_supervised_heatmap_multiview_transformer(self):
-        """Returns SemiSupervisedHeatmapTrackerMultiviewTransformer for semi-supervised variant."""
-        from lightning_pose.models import SemiSupervisedHeatmapTrackerMultiviewTransformer
-        assert (
-            get_model_class('heatmap_multiview_transformer', semi_supervised=True)
-            is SemiSupervisedHeatmapTrackerMultiviewTransformer
-        )
-
-    def test_get_model_class_semi_supervised_raises_for_unknown(self):
-        """Raises NotImplementedError for an unrecognised semi-supervised model_type."""
-        with pytest.raises(
-            NotImplementedError, match='invalid model_type for a semi-supervised',
-        ):
-            get_model_class('unknown_type', semi_supervised=True)
 
 
 def test_predict_dataset(cfg, heatmap_data_module, tmpdir):
